Ajoute les onglets de comparaison sur le test non réalisé également
